requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
transformers>=4.30.0
torch-geometric>=2.3.0
dgl>=1.1.0  # For graph operations

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
pillow>=9.5.0

# Training utilities
tqdm>=4.65.0
wandb>=0.15.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Development tools
pytest>=7.3.0
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0 